Fix tests that broke through renaming
